pyproject.toml
[project]
name = "datareader"
version = "0.1.0"
description = "A simple utility package for reading input data from text into a Python session or PostgreSQL database."
readme = "README.md"
packages = [{include = "datareader", from = "src"}]
authors = [
    { email = "sv973005@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "attrs>=25.1.0",
    "black>=25.1.0",
    "coverage-badge>=1.1.2",
    "pandas>=2.2.3",
    "pandas-stubs>=2.2.3.241126",
    "psycopg2-binary>=2.9.10",
    "pytest-cov>=6.0.0",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.9.10",
    "yamllint>=1.35.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88

[tool.mypy]
files = "."
disallow_any_generics = true
check_untyped_defs = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "datareader"
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = "--cov=datareader"
filterwarnings = []

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E", "F", "W", # pycodestyle + pyflakes == flake8 - mccabe
    "I",           # isort
    "B",           # flake8-bugbear
    "UP",          # pyupgrade
    "S",           # flake8-bandit
    "C4",          # flake8-comprehensions
    "PIE",         # flake8-pie
    "SIM",         # flake8-simplify
    "PTH",         # flake8-use-pathlib
]
ignore = [
    "E402",   # For example: src/datareader/loader/__init__.py
    "S101",   # We don't use `python -o`, so `assert` is safe.
    "PTH123", # Worth considering, but involves updates in some other places to keep mypy happy.
]
